requirements.txt
pillow>=9.0.0
pillow-heif>=0.7.0
piexif>=1.1.3
exifread>=3.0.0
